+++ server/src/main/java/org/allaymc/server/command/tree/node/PosNode.java
@@ -5,7 +5,7 @@ import org.allaymc.api.command.tree.CommandContext;
 import org.allaymc.api.command.tree.CommandNode;
 import org.cloudburstmc.protocol.bedrock.data.command.CommandParam;
 import org.cloudburstmc.protocol.bedrock.data.command.CommandParamData;
-import org.joml.Vector3f;
+import org.joml.Vector3d;
 
 /**
  * @author daoge_cmd
@@ -22,7 +22,7 @@ public class PosNode extends BaseNode {
         var baseY = basePos.y();
         var baseZ = basePos.z();
 
-        float x, y, z;
+        double x, y, z;
         try {
             var index = context.getCurrentArgIndex();
             x = getRelative(context.queryArg(index), baseX);
@@ -31,14 +31,14 @@ public class PosNode extends BaseNode {
         } catch (Exception e) {
             return false;
         }
-        context.putResult(new Vector3f(x, y, z));
+        context.putResult(new Vector3d(x, y, z));
         context.popArg();
         context.popArg();
         context.popArg();
         return true;
     }
 
-    protected float getRelative(String token, float base) {
+    protected double getRelative(String token, double base) {
         if (!token.contains("~")) return Float.parseFloat(token);
 
         if (token.length() > 1) return base + Float.parseFloat(token.substring(1));
