+++ server/src/main/java/org/allaymc/server/entity/component/event/CEntityFallEvent.java
@@ -10,5 +10,5 @@ import org.allaymc.api.eventbus.event.Event;
 @Getter
 @AllArgsConstructor
 public class CEntityFallEvent extends Event {
-    protected float fallDistance;
+    protected double fallDistance;
 }
